:art: rename views to routes
