+++ package.json
@@ -5,8 +5,9 @@
   "main": "app.js",
   "scripts": {
     "test": "npm run test",
-    "start": "nodemon app.js",
-    "lint": "eslint . --fix"
+    "start": "node app.js",
+    "lint": "eslint . --fix",
+    "dev": "nodemon app.js"
   },
   "repository": {
     "type": "git",
